imagelib index

Library imagelib

This library exposes the following toplevel modules:

Image

ImageBMP

ImageGIF

ImageJPG

ImageLib Note: This module is an interface to the pure OCaml implementations of various image formats.

ImagePNG

ImagePPM

ImageUtil

ImageXCF

Library imagelib.unix

This library exposes the following toplevel modules:

ImageLib_unix This module provides an easy-to-use interface for imagelib. In most cases, you'd want to use these functions rather than those in imagelib.

ImageUtil_unix